ref: Rework arrow spells to use forceMoveProjectile
