+++ src/cards/arrow_fork.ts
@@ -27,23 +27,20 @@ const spell: Spell = {
     animationPath: '',
     sfx: 'arrow',
     description: ['spell_arrow_fork', damageDone.toString()],
-    effect: arrowEffect(1, damageDone, fireForkedArrows)
+    effect: arrowEffect(1)
+  },
+  events: {
+    onProjectileCollision: ({ unit, underworld, projectile, prediction }) => {
+      if (unit) {
+        Unit.takeDamage(unit, damageDone, projectile.startPoint, underworld, prediction, undefined, { thinBloodLine: true });
+        // Now fork into regular arrows that fire in directions
+        for (let newAngle of [Math.PI / 12, -Math.PI / 12, 2 * Math.PI / 12, -2 * Math.PI / 12, 3 * Math.PI / 12, -3 * Math.PI / 12]) {
+          const angle = getAngleBetweenVec2s(projectile.startPoint, unit) + newAngle;
+          const castLocation = getEndpointOfMagnitudeAlongVector(unit, angle, 10_000);
+          arrowEffect(1)({ cardIds: [regularArrow.card.id], shouldRefundLastSpell: false, casterPositionAtTimeOfCast: unit, targetedUnits: [], targetedPickups: [], casterUnit: unit, castLocation, aggregator: { unitDamage: [], radius: 0 }, initialTargetedPickupId: undefined, initialTargetedUnitId: undefined }, regularArrow.card, 1, underworld, prediction, false);
+        }
+      }
+    }
   }
 };
-async function fireForkedArrows(state: EffectState, firstTarget: Unit.IUnit, underworld: Underworld, prediction: boolean) {
-  return raceTimeout(5_000, 'fireForkedArrows', new Promise((resolve) => {
-    let promises = [];
-    // Now fork into regular arrows that fire in directions
-    for (let newAngle of [Math.PI / 12, -Math.PI / 12, 2 * Math.PI / 12, -2 * Math.PI / 12, 3 * Math.PI / 12, -3 * Math.PI / 12]) {
-      const angle = getAngleBetweenVec2s(state.casterUnit, firstTarget) + newAngle;
-      const castLocation = getEndpointOfMagnitudeAlongVector(firstTarget, angle, 10_000);
-      // Override casterUnit as firstTarget so forked arrows don't hit the target that they are forking off of
-      promises.push(arrowEffect(1, 10, undefined, true)({ ...state, casterPositionAtTimeOfCast: firstTarget, targetedUnits: [], casterUnit: firstTarget, castLocation }, regularArrow.card, 1, underworld, prediction, false));
-    }
-    Promise.all(promises).then(() => {
-      resolve(state);
-    });
-  }));
-
-}
 export default spell;
