+++ src/cards/arrow_triple.ts
@@ -3,6 +3,7 @@ import { Spell } from './index';
 import { CardRarity, probabilityMap } from '../types/commonTypes';
 import { arrowEffect } from './arrow';
 import { arrow2CardId } from './arrow2';
+import { takeDamage } from '../entity/Unit';
 
 export const arrowTripleCardId = 'Triple Arrow';
 const damageDone = 10;
@@ -24,7 +25,14 @@ const spell: Spell = {
     animationPath: '',
     sfx: 'arrow',
     description: ['spell_arrow_many', arrowCount.toString(), damageDone.toString()],
-    effect: arrowEffect(arrowCount, damageDone)
+    effect: arrowEffect(arrowCount)
+  },
+  events: {
+    onProjectileCollision: ({ unit, underworld, projectile, prediction }) => {
+      if (unit) {
+        takeDamage(unit, damageDone, projectile.startPoint, underworld, prediction, undefined, { thinBloodLine: true });
+      }
+    }
   }
 };
 export default spell;
